Add more unit tests for github repo reader
- split the one test file into 3
- comphrensively tests github repo with different combination of filters.
